backend/requirements.txt
Django==2.1.5
django-choices==1.6.1
djangorestframework==3.9.0
djangorestframework-camel-case==0.2.0
drf-nested-routers==0.91
psycopg2-binary==2.7.5
factory_boy==2.11.1
Faker==0.9.2
